# swagger.yaml
openapi: 3.0.3
info:
  title: Finance Tracker API
  description: API для ведення особистого бюджету з функціями створення, редагування, видалення витрат, перегляду статистики та експорту даних у CSV.
  version: 1.0.0
servers:
  - url: http://localhost:3001
    description: Local server
paths:
  /api/expenses:
    post:
      summary: Створити нову витрату
      tags:
        - Expenses
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpenseInput'
      responses:
        '201':
          description: Витрата створена успішно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expense'
        '400':
          description: Неправильні дані (наприклад, категорія не знайдена)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Помилка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      summary: Отримати всі витрати
      tags:
        - Expenses
      responses:
        '200':
          description: Список всіх витрат
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Expense'
        '500':
          description: Помилка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/expenses/{id}:
    put:
      summary: Оновити витрату
      tags:
        - Expenses
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID витрати
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpenseInput'
      responses:
        '200':
          description: Витрата оновлена успішно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expense'
        '404':
          description: Витрату не знайдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Помилка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Видалити витрату
      tags:
        - Expenses
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID витрати
      responses:
        '200':
          description: Витрата видалена успішно
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Expense deleted successfully
        '404':
          description: Витрату не знайдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Помилка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/expenses/statistics:
    get:
      summary: Отримати статистику витрат
      tags:
        - Expenses
      parameters:
        - name: startDate
          in: query
          required: false
          schema:
            type: string
            format: date
          description: Початкова дата (формат YYYY-MM-DD)
        - name: endDate
          in: query
          required: false
          schema:
            type: string
            format: date
          description: Кінцева дата (формат YYYY-MM-DD)
      responses:
        '200':
          description: Статистика витрат
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Statistics'
        '500':
          description: Помилка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/expenses/export:
    get:
      summary: Експортувати витрати у CSV
      tags:
        - Expenses
      responses:
        '200':
          description: Файл CSV із витратами
          content:
            text/csv:
              schema:
                type: string
                format: binary
        '500':
          description: Помилка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/categories:
    post:
      summary: Створити нову категорію
      tags:
        - Categories
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryInput'
      responses:
        '201':
          description: Категорія створена успішно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '500':
          description: Помилка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      summary: Отримати список категорій
      tags:
        - Categories
      responses:
        '200':
          description: Список категорій
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
        '500':
          description: Помилка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/categories/{id}:
    put:
      summary: Оновити категорію
      tags:
        - Categories
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID категорії
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryInput'
      responses:
        '200':
          description: Категорія оновлена успішно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '404':
          description: Категорію не знайдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Помилка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Видалити категорію
      tags:
        - Categories
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID категорії
      responses:
        '200':
          description: Категорія видалена успішно
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Category deleted successfully
        '400':
          description: Категорія пов'язана з витратами
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Категорію не знайдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Помилка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    ExpenseInput:
      type: object
      required:
        - amount
        - category
      properties:
        amount:
          type: number
          example: 50
        category:
          type: string
          format: ObjectId
          example: 507f1f77bcf86cd799439011
        date:
          type: string
          format: date-time
          example: 2025-05-01T00:00:00Z
        description:
          type: string
          example: Lunch
    Expense:
      type: object
      properties:
        _id:
          type: string
          example: 507f1f77bcf86cd799439011
        amount:
          type: number
          example: 50
        category:
          type: object
          properties:
            _id:
              type: string
              example: 507f1f77bcf86cd799439011
            name:
              type: string
              example: Food
            description:
              type: string
              example: Food expenses
        date:
          type: string
          format: date-time
          example: 2025-05-01T00:00:00Z
        description:
          type: string
          example: Lunch
        createdAt:
          type: string
          format: date-time
          example: 2025-05-01T12:00:00Z
        updatedAt:
          type: string
          format: date-time
          example: 2025-05-01T12:00:00Z
    Statistics:
      type: object
      properties:
        totalAmount:
          type: number
          example: 80
        byCategory:
          type: array
          items:
            type: object
            properties:
              _id:
                type: string
                example: Food
              totalAmount:
                type: number
                example: 50
        averageAmount:
          type: number
          example: 40
    CategoryInput:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          example: Transport
        description:
          type: string
          example: Transportation costs
    Category:
      type: object
      properties:
        _id:
          type: string
          example: 507f1f77bcf86cd799439011
        name:
          type: string
          example: Transport
        description:
          type: string
          example: Transportation costs
        createdAt:
          type: string
          format: date-time
          example: 2025-05-01T12:00:00Z
        updatedAt:
          type: string
          format: date-time
          example: 2025-05-01T12:00:00Z
    Error:
      type: object
      properties:
        message:
          type: string
          example: Something went wrong!
        error:
          type: string
          example: Error message